refactor: update publish script and clean up type exports
- Added a `--tag smoke` option to the `pnpm publish` command in the publish-local-registry.sh script.
- Refactored the index.ts file to clean up exports, moving type exports to a dedicated section and removing unused types.
- Streamlined the useFireproof.ts file by consolidating imports and removing deprecated code, enhancing clarity and maintainability.
