+++ src/react/index.ts
@@ -1,3 +1,15 @@
-export { FireproofCtx, type UseFireproof, useFireproof, type LiveQueryResult, type UseDocumentResult } from "./useFireproof.js";
+export { FireproofCtx, useFireproof } from "./useFireproof.js";
+
+export type {
+  UseFireproof,
+  LiveQueryResult,
+  UseDocumentResult,
+  AllDocsResult,
+  ChangesResult,
+  UseDocument,
+  UseLiveQuery,
+  UseAllDocs,
+  UseChanges,
+} from "./types.js";
 
 export * from "./img-file.js";
